stat: improve memory when to use stat cache
