+++ statistics/handle/cache/internal/inner.go
@@ -50,4 +50,5 @@ type StatsCacheInner interface {
 	// Front returns the front element's owner tableID, only used for test
 	// TODO: this method is mainly for test, remove it in the future.
 	Front() int64
+	Release()
 }
