+++ statistics/handle/cache/internal/mapcache/BUILD.bazel
@@ -2,7 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "mapcache",
-    srcs = ["map_cache.go"],
+    srcs = [
+        "sharding_map_cache.go",
+        "table_cache.go",
+        "table_cache_view.go",
+    ],
     importpath = "github.com/pingcap/tidb/statistics/handle/cache/internal/mapcache",
     visibility = ["//statistics/handle:__subpackages__"],
     deps = [
